Part3 refactoring step01

+++ src/main/java/org/stepup/stream6/interfaces/CurrRuleAble.java
@@ -1,7 +1,6 @@
 package org.stepup.stream6.interfaces;
 
-import java.math.BigDecimal;
 @FunctionalInterface
 public interface CurrRuleAble {
-    boolean check(BigDecimal bgnum);
+    boolean check(Integer bgnum);
 }
